+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog
 
 All notable changes to `charlielangridge/lunar-xero` will be documented in this file.
+
+## v0.5.0 - 2026-05-08
+
+- Added support for syncing and emailing authorised Xero invoices.
+- Added Xero invoice `sent_to_contact` checks to avoid duplicate customer invoice emails.